requirements.txt
APScheduler==3.6.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.3
six==1.13.0
SQLAlchemy==1.3.11
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
